refactor(test): 인수테스트코드 리팩토링 RestAssured 메서드 분리
